+++ sc/social/like/setuphandlers.py
@@ -1,4 +1,27 @@
+# -*- coding:utf-8 -*-
 from Products.CMFCore.utils import getToolByName
+from Products.CMFQuickInstallerTool import interfaces as qi_interfaces
+from Products.CMFPlone import interfaces as plone_interfaces
+from zope.interface import implements
+
+
+class HiddenProducts(object):
+    implements(qi_interfaces.INonInstallable)
+
+    def getNonInstallableProducts(self):
+        return ['sc.social.like.upgrades.v2000',
+                'sc.social.like.upgrades.v3000']
+
+
+class HiddenProfiles(object):
+    implements(plone_interfaces.INonInstallable)
+
+    def getNonInstallableProfiles(self):
+        return [
+            u'sc.social.like:uninstall',
+            u'sc.social.like.upgrades.v2000:default',
+            u'sc.social.like.upgrades.v3000:default',
+        ]
 
 
 def install(context):
@@ -31,15 +54,3 @@ def uninstall(context):
             pp.manage_delObjects(ids='sc_social_likes_properties')
     except KeyError:
         pass
-
-
-def upgradefrom1001(context):
-    ''' Upgrade from 1001
-    '''
-    setup = getToolByName(context, 'portal_setup')
-
-    # Install
-    profiles = ['profile-sc.social.like:to2000', ]
-
-    for profile in profiles:
-        setup.runAllImportStepsFromProfile(profile)