+++ sc/social/like/testing.py
@@ -1,9 +1,44 @@
 # -*- coding: utf-8 -*-
-
+from PIL import Image
 from plone.app.testing import PloneSandboxLayer
 from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import IntegrationTesting
 from plone.app.testing import FunctionalTesting
+from StringIO import StringIO
+
+import random
+
+
+def generate_image(width, height):
+    # Mandelbrot fractal
+    # FB - 201003254
+    # drawing area
+    xa = -2.0
+    xb = 1.0
+    ya = -1.5
+    yb = 1.5
+    maxIt = 25  # max iterations allowed
+    # image size
+    image = Image.new("RGB", (width, height))
+    c = complex(random.random() * 2.0 - 1.0, random.random() - 0.5)
+
+    for y in range(height):
+        zy = y * (yb - ya) / (height - 1) + ya
+        for x in range(width):
+            zx = x * (xb - xa) / (width - 1) + xa
+            z = complex(zx, zy)
+            for i in range(maxIt):
+                if abs(z) > 2.0:
+                    break
+                z = z * z + c
+            r = i % 4 * 64
+            g = i % 8 * 32
+            b = i % 16 * 16
+            image.putpixel((x, y), b * 65536 + g * 256 + r)
+
+    output = StringIO()
+    image.save(output, format="PNG")
+    return output.getvalue()
 
 
 class Fixture(PloneSandboxLayer):